AzureFunction/Voice_Insight_Processor/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai==1.6.0
tiktoken
pandas
matplotlib
seaborn
pyodbc
python-dotenv
pydub
azure-storage-blob
azure.storage.blob
docker